Fixed issue trying to save company but needs refactor later§
