+++ src/services/companyservice.js
@@ -1,7 +1,6 @@
 /* eslint camelcase: 0 */
 const Q = require('q')
 const winston = require('winston')
-const metadataRepository = require('../repositorys/metadatarepository')
 const advisorRepository = require('../repositorys/advisorrepository')
 const companyRepository = require('../repositorys/companyrepository')
 const serviceDeliveryRepository = require('../repositorys/servicedeliveryrepository')
@@ -76,8 +75,6 @@ function getInflatedDitCompany (token, id) {
   })
 }
 
-
-
 function getCompanyForSource (token, id, source) {
   return new Promise((resolve, reject) => {
     Q.spawn(function *() {
@@ -103,36 +100,4 @@ function getCompanyForSource (token, id, source) {
   })
 }
 
-function setCHDefaults (token, company) {
-  return new Promise((resolve, reject) => {
-    Q.spawn(function *() {
-      try {
-        if (company.company_number) {
-          const ch = yield companyRepository.getCHCompany(token, company.company_number)
-          if (!company.name) company.name = ch.name
-          if (!company.registered_address_1) company.registered_address_1 = ch.registered_address_1
-          if (!company.registered_address_2) company.registered_address_2 = ch.registered_address_2
-          if (!company.registered_address_town) company.registered_address_town = ch.registered_address_town
-          if (!company.registered_address_county) company.registered_address_county = ch.registered_address_county
-          if (!company.registered_address_postcode) company.registered_address_postcode = ch.registered_address_postcode
-          if (!company.registered_address_country) company.registered_address_country = ch.registered_address_country.id
-          company.uk_based = true
-
-          // Business type
-          const businessTypes = metadataRepository.TYPES_OF_BUSINESS
-          for (const businessType of businessTypes) {
-            if (businessType.name.toLowerCase() === ch.company_category.toLowerCase()) {
-              company.business_type = businessType.id
-            }
-          }
-        }
-        resolve(company)
-      } catch (error) {
-        winston.error(error)
-        reject(error)
-      }
-    })
-  })
-}
-
-module.exports = { getInflatedDitCompany, getCompanyForSource, setCHDefaults }
+module.exports = { getInflatedDitCompany, getCompanyForSource }
